- Change LocationStrategy to HashLocationStrategy - add production environment (new task "start-prod") - rename sessionState files to dataState
