+++ BugTracker.App/Static/devIndex.html
